refactor: clean code and add more test case (#1552)
* refactor: clean code and add more test case

* fix: inventory crash and take creative item

* fix: boat offset

* fix: horse and boat offset and boundbox

* fix: pick slab

* fix: if offhand is null and it will not save

* fix: cactus place

* fix: try fix commit badge

* docs: commit autogenerated JaCoCo coverage badge

* fix: armors damage show

* fix: boat place

* fix: bucket place

* fix: bucket use

* fix: test

---------

+++ src/main/java/cn/nukkit/network/protocol/PlayerArmorDamagePacket.java
@@ -33,12 +33,11 @@ public class PlayerArmorDamagePacket extends DataPacket {
 
     @Override
     public void encode(HandleByteBuf byteBuf) {
-        
         int outflags = 0;
         for (PlayerArmorDamageFlag flag : this.flags) {
             outflags |= 1 << flag.ordinal();
         }
-        byteBuf.writeByte((byte) outflags);
+        byteBuf.writeByte(outflags);
 
         for (PlayerArmorDamageFlag flag : this.flags) {
             byteBuf.writeVarInt(this.damage[flag.ordinal()]);
